tradX/requirements.txt
python-binance
TA-Lib
numpy
websocket_client
telethon